# apps/demo-services/settlement-risk-spring/src/main/java/com/demo/risk/controller/CounterpartyRiskController.java
package com.demo.risk.controller;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import com.demo.risk.model.BreachRecord;
import com.demo.risk.model.BreachSearchRequest;
import com.demo.risk.model.CounterpartyLimitResponse;
import java.util.List;

@RestController
@RequestMapping("/risk")
public class CounterpartyRiskController {

    @GetMapping("/counterparties")
    public ResponseEntity<String> listCounterparties() {
        return ResponseEntity.ok("[]");
    }

    @GetMapping("/counterparties/{counterpartyId}")
    public ResponseEntity<String> getCounterparty(@PathVariable String counterpartyId) {
        return ResponseEntity.ok("{}");
    }

    @GetMapping("/limits/{counterpartyId}")
    public ResponseEntity<CounterpartyLimitResponse> getCounterpartyLimit(@PathVariable String counterpartyId) {
        return ResponseEntity.ok(null);
    }

    @GetMapping("/exposures/{counterpartyId}")
    public ResponseEntity<String> getCounterpartyExposure(@PathVariable String counterpartyId) {
        return ResponseEntity.ok("{}");
    }

    @PostMapping("/breaches/search")
    public ResponseEntity<List<BreachRecord>> searchBreaches(@RequestBody(required = false) BreachSearchRequest body) {
        return ResponseEntity.ok(List.of());
    }
}