# apps/demo-services/settlement-risk-spring/src/main/java/com/demo/risk/model/BreachSearchRequest.java
package com.demo.risk.model;

public class BreachSearchRequest {
    private String counterpartyId;
    private String severity;
    private String fromDate;

    public String getCounterpartyId() { return counterpartyId; }
    public void setCounterpartyId(String counterpartyId) { this.counterpartyId = counterpartyId; }
    public String getSeverity() { return severity; }
    public void setSeverity(String severity) { this.severity = severity; }
    public String getFromDate() { return fromDate; }
    public void setFromDate(String fromDate) { this.fromDate = fromDate; }
}